Correct pwdutils build recipe
Override do_configure() and do_compile() in pwdutils-extraconf.inc.
Using the generated recipe (pwdutils.inc) there is an issue during the
configure process. The origin of this issue is still unknown so, instead
of calling autotools_do_configure we manualy call autoreconf and
./configure which does the trick. Finding the origin of this issue and a
better solution would be appreciable.

Some files are also not added in the package using the generated recipe,
also for an unknown reason which is very likely to be found in the
spec2yocto recipe generator. Those files are manually added in
pwdutils-extraconf.inc.

Change-Id: I7fea88cba28aa5c7604bc0d20975d65808777d3d
Signed-off-by: Kévin THIERRY <kevin.thierry@open.eurogiciel.org>