Display desktop background
Since last desktop-skin recipe update
4b8829828a9eaa019101e05ec3bb97000b12a0bf desktop wallpaper wasn't
displayed anymore. This update replaced the hard-coded image name in
weston.ini with a symbolic link to the image file.

In weston-common-extraconf.inc it was possible to define a different
background, now this is done in desktop-skin since the symbolic link is
made in this package.

Bug-Tizen: BTY-6
Change-Id: I74f88ff574d38bc3e5f7c855a8d6048197e957a3
Signed-off-by: Kévin THIERRY <kevin.thierry@open.eurogiciel.org>