Use the Tizen Common kernel configuration
A lot of options were missing in the default configuration. To get a
kernel configuration very similar to the one used in Tizen Common, two
things are done:
* Setting the "KCONFIG_MODE" variable to prevent the use of the
  "--allnoconfig" flag which disable all kernel options.
* Use the defconfig files from the Tizen Common kernel (project
  kernel-common).

The difference between the Tizen Common kernel configuration and the
Tizen Yocto one are now minimal. There are due to two things:
* Use of kernel 3.14.5 in Yocto instead of 3.14.4.
* Additional BSP/MACHINE specific and Yocto specific configurations.

Also rename the subdir for additional files "linux-yocto".

Change-Id: Id9dd111975ff0e5deaa84567aaeeb61908c192b2
Signed-off-by: Kévin THIERRY <kevin.thierry@open.eurogiciel.org>