vconf: explicitly add dependencies
This prevents vconf build from failing. It is not clear why an explicit
dependency to glib-2.0 is required for the target build since it this
dependency is already present in the recipe (for both native and target
builds).

This commit is linked to commit
c0da18eab9765139f6b44f10ea926e7c2a87ebe2.

Bug-Tizen: BTY-14
Change-Id: Iacada33b12137a319021a771fce00a04d6899c17
Signed-off-by: Kévin THIERRY <kevin.thierry@open.eurogiciel.org>