Add distro conf
Tizen now uses its own distro conf (based on the default one) instead
of only relying on the dfault one.

The file tcmode-tizen.inc is introduced to replace the file
tcmode-default.inc in order to correct a conflict between glibc-initial
and nativesdk-glibc-initial. If a better solution is found to solve this
conflict, please remove this file and set TCMODE to "default" in
tizen.conf.

Change-Id: I91d5e04f3fe12e596a21cfe4ee563c44aaeae72c
Signed-off-by: Kévin THIERRY <kevin.thierry@open.eurogiciel.org>