Copy manifest file into ${PKGDEST}
If running do_package_write_rpm with a build from sstate, it can rely on
the output of do_package being present, it cannot rely on the presence
of ${S} any longer. This patch copyies the manifests into ${PKGDEST}
which is stored output of do_package and hence preserved, fixing the
build failure you'd otherwise see if you force a build of
do_package_write_rpm from sstate.

[kevin.thierry: add manifest.bbclass to automatically copy the manifest
when a recipe inherits manifest.]

Bug-Tizen: BTY-59
Change-Id: I34f6d3ec53e40d64958da8587501380e47121729
Signed-off-by: Kévin THIERRY <kevin.thierry@open.eurogiciel.org>