manifests: add support for SRC_URI
MANIFESTFILES assumes that the file comes out of the source or build
process. The new SRC_URI support for files ending in .manifest or
.manifest.in makes it possible to add manifest files as pure packaging
meta data.

The "pkg" parameter determines which package the manifest file is used
for. If not set or empty, it is used for all. If set, the value is a
comma-separated list of package names. The first suitable manifest
file is used. Fake example file_%.bbappend:

SRC_URI += "file://libmagic.manifest.in;pkg=libmagic,libmagic-data file://file.manifest"
FILESEXTRAPATHS_prepend := "${THISDIR}/${PN}:"

When turning a .manifest.in file into .manifest, @[a-zA-Z0-9]+@ will
get replaced by the value of the corresponding Bitbake variable (for
example, "bindir").

This could also cover Tizen multi-user variables
(https://wiki.tizen.org/wiki/Multi-user_Platform_Metadata) if those were
added to the active distro or build conf.

Change-Id: I33e49ea4e97670a70bc06bff21a265272a48626e
Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>