manifests: default Tizen manifest
All binary .rpms should be built with at least a stub .manifest file
assiging the "_" label to all packaged files. Explicitly adding that
to all Yocto recipes via .bbappend is cumbersome, so do it globally in
tizen.conf. It's still possible to override that default by prepending
a different .manifest[.in] file.

Change-Id: Ife3aa29c42c0e095f913c9b6effae78e5caa7e10
Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>