tizen.conf: use Tizen systemd_unitdir
Tizen uses a different path for systemd unit files than Yocto. Using
the path from Tizen makes it possible to use packages from Tizen
without having to modify them first.

The need for this in .spec files will go away with the next update
of the translator tool, because then it'll use the rpm macro.

Whether it's relevant for Yocto-based builds depends on how fixed
paths in Tizen packages are handled - not sure about that.

Change-Id: I366d359ed2ab7654230090eab5e3be29121acf35