connman: fix dependencies
Yocto lists dependencies which are not used in Tizen. Probably they
are needed for features which were considered irrelevant for Tizen.
Therefore remove them for builds *and* conversion to .spec.

Also, a new dependency on readline caused by one of the Tizen
patches was not specified. Probably worked with Yocto accidentally
(readline installed because something else depended on it).

Change-Id: Ia7c3f733c6abf8aa447097f2f3f940e279cc7796