libpkg: utility class for defining lib packages
There's a lot of repetitive work which needs to be done when putting
shared libraries into their own package: define the package, make the
-dev package depend on it, enabled ldconfig in .spec.

This class automates all that. To define a lib${PN} package, one only
needs to "inherit libpkg" in a .bb or .bbappend. To define one package
per library, set PKG_LIBS. See the class for details.

The new packages get created when building binaries and when converting
to .spec.

Change-Id: I30323976b5baaeda15913702da86794fadeb2198