libpkg: support augmenting existing lib packages
In the poppler recipe, ${PN}-dev does not depend on lib${PN} (upstream
bug?!) and SRPM_IS_LIB does not contain lib${PN} (because that's
Tizen-specific). Listing lib${PN} in addition to new packages in
PKG_LIBS is an easy way of fixing this.

However, the old package should better not be listed twice in
PACKAGES, because other code (for example, the meta translator) do not
expect that.

Change-Id: Ifec1eb2707fe434333daec91d0ee158aef8e1297