desktop-file-utils: avoid host pollution
Building fails when Emacs is installed on the host. The reason
is that the .spec file assumes that Emacs is not installed and
installing the lisp files needs to be done manually after the
desktop-file-utils "make install".

The actual failure then is:
...
|   test ! -f /data/yocto/build/tizen-distro/ivi/x86_64/tmp-glibc/work/core2-64-oe-linux/desktop-file-utils/0.20-r0/image/usr/share/emacs/site-lisp/desktop-entry-mode.el
| ERROR: Function failed: do_install (log file is located at
/data/yocto/build/tizen-distro/ivi/x86_64/tmp-glibc/work/core2-64-oe-linux/desktop-file-utils/0.20-r0/temp/log.do_install.14197)
ERROR: Task 6589 (/data/yocto/tizen-distro/meta-tizen/meta-tizen-common-base/recipes-core/desktop-file-utils/desktop-file-utils_git.bb,do_install) failed with exit code '1'

To avoid the check for Emacs, pre-set the autoconf variable.

Change-Id: Ib96451edf8b5ce13f9e957668b4406bd3e64209d
Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>