Tizen recipes: avoid -j16
-j16 is too high for users on smaller machines, leading to swapping and
out-of-memory situations, including random killing of processes by the
kernel's OOM killer.

This commit is the result of a global search/replace of -j16 with
${PARALLEL_MAKE}. It touches files generated with spec2yocto, so beware to fix
that tool before refreshing recipes.

In some cases (like calling oe_runmake), passing -j is redundant because it is
already included in the default parameters. This commit does not try to
address that, so one may see things like (from ail):

+ oe_runmake -j 16
+ oe_runmake_call -j 16
+ bbnote make -j 16 -j 16
+ echo 'NOTE: make -j 16 -j 16'
NOTE: make -j 16 -j 16

The "-j 16" here corresponds to ${PARALLEL_MAKE}.

Change-Id: Id7d7285a4ae88fdb44292758de79d98537e12f58
Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>