vala: fix valac circular dependency
Building vala-native randomly failed to build depending on the order
in which files were checked out by git: if the .stamp files happened
to be older than the .vala files, valac would be called to regenerate
the .c files, but that fails without valac.

This would not happen when untarring the pristine upstream source,
but spec2yocto does not support that mode.

The Fix-time-stamp.patch may have been meant to address that, but despite
refreshing it, builds still failed. The new approach is to fix up time stamps
differently for native (do not rebuild source) and target (always rebuild).

Change-Id: I3b8f0e7b275d7032ec460ea15e5f1d1beeec518d