libmm-log: manually fix and update
spec2yocto was tracking an obsolete revision of the source via
manifest_fix.xml. That revision was too old for libmm-debug, which caused
compile errors.

For now, update to the later version manually, ignoring the missing manifest
file.

That Samsung's update of the upstream code dropped the explicit manifest file
is not nice, but should not cause issues because both OBS and "Tizen on Yocto"
have default manifests (see d47ae747b "manifests: default Tizen manifest").

Change-Id: I2bdb87a52b9f899730aecd325c02a7cd50942ead