automotive-message-broker.inc: manually disable QT5 dependency
Some of the new dependencies are wrong for the Bitbake Qt recipes.
Instead of opening that can of worms now, better keep these optional
features disabled.

The right solution will be to update the spec2yocto rpm build config.
Doing it manually is only the first step to get going again.

Change-Id: Ie3f2dd9ac050f6e32a8717ad7d3ebe4585715ba5