genivi-shell: update build fixes
The recent spec2yocto refresh led to different build rules.
The wayland-scanner invocation is now in the .inc file and needs
to be fixed there.

We also must ensure that it happens as part of do_compile,
because the wayland-native dependency is only guaranteed to
be resolved for that task and not the earlier do_patch.

Change-Id: I5890e7a47d96799eceb7f5a425386af8e6b0a436