weston-ivi-shell-extraconf.inc: also remove wayland.png
These files need to be removed because they overlap with
weston. The .spec file does it by excluding them from
the files list, which is something that does not work
when building with Yocto.

A better solution in the .spec file would be to delete
the files after %install, which would work when translating
to Yocto.

Change-Id: If2a9dec2a18e2ed596161a88588314e233fb6084