Tizen IVI image: remove unused Common packages
This removes wrt-widgets, user-session-units and meta-common,
which are all not used in Tizen IVI 3.0 and thus were also
removed from the current recipe set.

They are intentionally left in the Common image configuration
because they might server a purpose there, even if those
configurations are now unusable as they stand.

Change-Id: I62e61a703f6d2fd027cdb971976230c419107d0b