crosswalk-extraconf.inc: disable crosswalk-dbg package
When enabled, it gets so large (?) that the "installed"
size entry in the package feed turns into an invalid "None".
To work around this tools issue, disable the package.

This leads to QA warnings about installed but not packaged
files. This needs to be solved separately, ideally by not having
these files generated in the first place.

Change-Id: I991604395839a430ecc7a8fc878d75728ce63cad