opencv: bump version and disable ffmpeg to match Tizen
The meta-oe opencv_2.4.bb no longer compiles, something
related to ffmpeg support.

That is not enabled in Tizen anyway, so we could use that
recipe and just modify the PACKAGECONFIG in a .bbappend file.
But as Tizen is on a different, more recent version, better
copy the recipe and update it to use source and default
features as in Tizen.

Change-Id: I8bdd14b5e2d57399e7bd8b6fcc4525f0ad32fa81