tlm-extraconf.inc: remove hacks, disable ico
The old hacks for getting IVI configuration packages created
are no longer necessary, because the base .inc file is now
the one for IVI.

Unfortunately building tlm-config-ivi-singleseat-ico runs
into a whole range of issues because of the dependency on
a native version of the pkgmgr. Even if we could it get to
build, whether it would work is another question. Ignoring
this issue for now...

Change-Id: I95190032007198702c394528116e2bb08ea03987