tlm.inc: manually fix recipe
spec2yocto seems to have problem with multiple postinst sections: they were
all added to the main package, thus making it fairly random which tlm.conf
symlink was really created.

This patch also fixes an upstream issue: the original .spec file packages the
same file multiple times in different packages.  We cannot do that with OE,
because a file only gets added to the first package that matches it. Replaced
with a shared tlm-config-ivi-common package.

Change-Id: I273b6da2fcfe3ca3ad14ae33af1c22c9e2fd3867