base-files: update fstab
The "tizen_fstab" was the same as the original OE fstab; installing
it instead of the original file is pointless.

But we *do* need to modify the original fstab to solve one issue:
systemd-fsck-root.service looks at the file system entry and runs
"fsck.auto" if we keep OE's "auto" entry.

This does not prevent booting, only the file system checking part
of it.

To replace the original file, we use the same name because then
the modified file will be used even when we revert the do_install
override.