config-tizen-ivi: add users to "video" and "input" group
This is important for launching weston: otherwise it fails to
read /dev/dri resp. /dev/input. This is different from normal
Tizen where users are part of "video", but not "input". Instead
weston-launch is suid root.

"weston-launch" no longer seems to be used at all. Perhaps it
was used to control access to the suid root weston-launch.

Needs further investigation and consolidation with Tizen...

Change-Id: I7aadbcc758434527192d3725bd9c1462f84c3a7a