filesystem layout: configure via base-files instead of config-image
Changing file permissions via postinst is bad style, for example,
rpm cannot be used to determine why permissions were set this way.
The right way of solving this is via rpm manifest files.

The whole custom config-image can be replaced with a modified base-file
that has a manifest file and DIRFILES_${PN} set up accordingly.

In adddition, this commit also makes "Tizen on Yocto" more like Tizen:
- no more /var/volation - that dependend on a systemd unit from OE
  that was not even present in "Tizen on Yocto"
- /sbin /lib and /bin are now symlinks to /usr

The advantage of these symlinks are that recipes from Tizen which
contain hard-coded paths will work without having to be fixed. It is
uncertain whether there *are* any such recipes at the moment, but
there might be some or might be in the future.

The real advantage right now is that it makes it simpler to compare
file systems from Tizen and "Tizen on Yocto".

Change-Id: I2a1dbdc6de58cdfd245a3c42b6f7a0f8855cb97d