config-tizen-ivi, weston: remove users from "input" group, suid root
It turns out that the suid root bit is needed for weston-launch after
all. It was also set in "Tizen on Yocto" builds published on the
Wiki, but it remains unclear how that was achieved.

The bit was missing because spec2yocto does not translate the permission part
of the file list in the .spec file. We need to add that back manually.