smack setup: define rules and set labels
The Tizen three-domain-model rules get created as part of the base-files.
They were taken from a live Tizen system. It remains unclear where they
are maintained in Tizen because the file was not owned by any package.

Setting the corresponding Smack labels and types has to be split between
base-files (regular directories in the filesystem) and
systemd-tmpfiles-setup.service (for directories created by systemd).

Setting Smack labels is done by calling chsmack directly. This avoids
the dependency on a patched rpm with manifest support and is also
the direction into which Tizen is going.

Change-Id: I897415375002a9e49c0a3805ac71d7eeba4f5f40