systemd: add systemd recipes for 2.16
The Tizen patches were meant to be used with systemd 2.16. OE-core
master no longer has that version, so we need to add it back in
the layer. Long term we need to move the patches to the version
used by OE-core, to benefit from systemd enhancements and fixes
developed upstream and/or OE-core.

This particular version of the systemd recipes comes from OE-core master
e9b13a which is the last revision which still had it.

Change-Id: I030ec5c9f226a5715f0e4349eda1d50b96eb458e