layer: introduce smack override
Making the meta data changes depend on a new "smack" override has the
desirable effect that one can include the layer in a bblayers.conf and then
decide separately whether Smack support is really enabled.

The goal is to do that via DISTROOVERRIDES in a distro conf, potentially
depending on a distro feature.

To avoid override name conflicts, the smack recipe had to be
renamed. "smack-userspace" was chosen, with "smack" being kept as alias.

While at it, replace the auto-generated recipe with a manual one building the
same source. Perhaps this should even pull directly from the upstream github
repo?

Change-Id: I0cada48bd4500f32e7624ac243992252bb856e7c